Move GameInfo to separate file
