+++ index.html
@@ -89,6 +89,7 @@
 
                     "child_shape": "./scripts/child_shape.js",
                     "editor": "./scripts/editor.js",
+                    "game_info": "./scripts/game_info.js",
                     "rigid_body": "./scripts/rigid_body.js",
                     "utils": "./scripts/utils.js"
                 }
