+++ scripts/main.js
@@ -6,126 +6,6 @@ import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
 import { ChildShape } from "child_shape";
 import { editor } from "editor";
 
-// Classes
-
-let Game;
-
-class gameInfo {
-    addMod(fileId, localId) {
-        this.modList.push({ fileId: fileId, localId: localId });
-
-        info_mods.textContent = this.modList.length;
-
-
-        this.addModToTable({ fileId: fileId, localId: localId }, this.modList.length-1);
-    }
-
-    addModToTable(mod, position) {
-        let currentRow = table_mods.insertRow(position+1);
-        let fileId_Cell = currentRow.insertCell(0);
-        let localId_Cell = currentRow.insertCell(1);
-        let button_remove_Cell = currentRow.insertCell(2);
-
-
-        var linkElement = document.createElement("a");
-        linkElement.innerText = mod.fileId;
-        linkElement.href = "https://steamcommunity.com/sharedfiles/filedetails/?id="+mod.fileId;
-        linkElement.target = "_blank";
-        linkElement.classList.add("link_alterantive_1");
-        fileId_Cell.appendChild(linkElement);
-
-        localId_Cell.textContent = mod.localId;
-
-        var buttonElement = document.createElement("button");
-        buttonElement.innerText = "🗙";
-        buttonElement.classList.add("table_remove");
-        button_remove_Cell.appendChild(buttonElement);
-
-        buttonElement.addEventListener("click", () => {
-            let modID = currentRow.rowIndex-1;
-            this.modList.splice(modID,1);
-            currentRow.remove();
-            info_mods.textContent = this.modList.length;
-        });
-    }
-
-    constructor(data) {
-        this.savegameversion = data[0];
-        this.flags = data[1];
-        this.seed = data[2];
-        this.gametick = data[3];
-        this.mods = data[4];
-        this.uniqueIds = data[5];
-
-        console.log(data);
-        this.ChildShapeID = data[5][15] + (data[5][14]<<8) + (data[5][13]<<16) + (data[5][12]<<24);
-
-        this.modList = Array((this.mods[0]<<24)+(this.mods[1]<<16)+(this.mods[2]<<8)+this.mods[3]);
-
-        info_mods.textContent = this.modList.length;
-
-        for (let i = 0; i < this.modList.length; i++) {
-            let fileId = 0;
-            for (let j = 0; j < 8; j++) {
-                fileId = fileId << 8;
-                fileId+= this.mods[i*25+4+j];
-            }
-
-            let localId = readUUID(this.mods, i*25+27);
-
-            this.modList[i] = { fileId: fileId, localId: localId };
-
-            this.addModToTable(this.modList[i], i);
-        }
-    }
-
-    advanceChildShapeID() {
-        this.ChildShapeID++;
-        this.uniqueIds[15] = this.ChildShapeID%256;
-        this.uniqueIds[14] = (this.ChildShapeID>>8)%256;
-        this.uniqueIds[13] = (this.ChildShapeID>>16)%256;
-        this.uniqueIds[12] = (this.ChildShapeID>>24)%256;
-
-        let statement = db.prepare("UPDATE Game SET uniqueIds = ?;");
-        statement.run([this.uniqueIds]);
-    }
-
-    updateDatabase() {
-        let statement = db.prepare("UPDATE Game SET savegameversion = ?, seed = ?, gametick = ?;");
-        statement.run([this.savegameversion, this.seed, this.gametick]);
-
-        let modData = new Uint8Array(this.modList.length*24+4);
-
-        modData[3] = this.modList.length%256;
-        modData[2] = (this.modList.length>>8)%256;
-        modData[1] = (this.modList.length>>16)%256;
-        modData[0] = (this.modList.length>>24)%256;
-
-
-        for (let i = 0; i < this.modList.length; i++) {
-            let thisFileId = this.modList[i].fileId;
-            for (let j = 0; j < 8; j++) {
-                modData[i*24+11-j] = thisFileId%256;
-                thisFileId = thisFileId >> 8;
-            }
-
-            let k = 0;
-            let UUID_position = i*24+27;
-            while (k < 36) {
-                modData[UUID_position--]=parseInt(this.modList[i].localId[k]+this.modList[i].localId[k+1], 16);
-                if (k==6||k==11||k==16||k==21) k+=3;
-                else k+=2;
-            }
-        }
-
-
-        this.mods = modData;
-
-        statement = db.prepare("UPDATE Game SET mods = ?;");
-        statement.run([modData]);
-    }
-}
-
 // Functions
 
 function select(type, objectID) {
@@ -605,8 +485,6 @@ open_file_button.onchange = () => {
         scene.add(light);
 
         editor.afterSaveLoad(r);
-
-        Game = new gameInfo(gameData);
 
         const infoGameVersion = document.getElementById("info_gameversion");
         const infoSeed = document.getElementById("info_gameSeed");
